+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # HEAD
 
+* Add `mousedown` and `mouseup` streams.
 * Change `createElementWithClass` to take type of node as an argument.
 
 # 0.1.3